fix(ProfitCalculator): 오버플로우를 막기 위해 long을 사용하라

+++ src/main/java/lotto/utility/ProfitCalculator.java
@@ -6,11 +6,11 @@ import java.util.Map;
 
 public class ProfitCalculator {
     public static String calculate(int purchaseCost, Map<String, Integer> matchedCount) {
-        long revenue = matchedCount.get(MatchedCountNameEnum.THREE_MATCHED.getMessage()) * LottoPrizeEnum.THREE_MATCHED_PRIZE.getAmount()
-                + matchedCount.get(MatchedCountNameEnum.FOUR_MATCHED.getMessage()) * LottoPrizeEnum.FOUR_MATCHED_PRIZE.getAmount()
-                + matchedCount.get(MatchedCountNameEnum.FIVE_MATCHED.getMessage()) * LottoPrizeEnum.FIVE_MATCHED_PRIZE.getAmount()
-                + matchedCount.get(MatchedCountNameEnum.FIVE_WITH_BONUS_MATCHED.getMessage()) * LottoPrizeEnum.FIVE_WITH_BONUS_MATCHED_PRIZE.getAmount()
-                + matchedCount.get(MatchedCountNameEnum.SIX_MATCHED.getMessage()) * LottoPrizeEnum.SIX_MATCHED_PRIZE.getAmount();
+        long revenue = (long)matchedCount.get(MatchedCountNameEnum.THREE_MATCHED.getMessage()) * LottoPrizeEnum.THREE_MATCHED_PRIZE.getAmount()
+                + (long)matchedCount.get(MatchedCountNameEnum.FOUR_MATCHED.getMessage()) * LottoPrizeEnum.FOUR_MATCHED_PRIZE.getAmount()
+                + (long)matchedCount.get(MatchedCountNameEnum.FIVE_MATCHED.getMessage()) * LottoPrizeEnum.FIVE_MATCHED_PRIZE.getAmount()
+                + (long)matchedCount.get(MatchedCountNameEnum.FIVE_WITH_BONUS_MATCHED.getMessage()) * LottoPrizeEnum.FIVE_WITH_BONUS_MATCHED_PRIZE.getAmount()
+                + (long)matchedCount.get(MatchedCountNameEnum.SIX_MATCHED.getMessage()) * LottoPrizeEnum.SIX_MATCHED_PRIZE.getAmount();
 
         double profit = (revenue / (double)purchaseCost) * 100;
         return new BigDecimal(profit).setScale(1, RoundingMode.HALF_UP).toPlainString();
